also move certificate unit tests into subdir
